Implementing paging3 for getPopularMovies
Still need to fix PopularMoviesRemoteMediator, LoadType.APPEND in loop. Need to discover why state.pages.data is not returning the last movie for the page.

+++ app/src/main/java/br/com/popularmovies/moviedetail/viewmodel/MovieDetailViewModel.kt
@@ -16,12 +16,12 @@ import kotlinx.coroutines.launch
 class MovieDetailViewModel @AssistedInject constructor(
     private val getMovieUseCase: GetMovieUseCase,
     private val saveMovieToFavoritesUseCase: SaveMovieToFavoritesUseCase,
-    @Assisted private val movieId: Int
+    @Assisted private val movieId: Long
 ) : ViewModel() {
 
     @AssistedInject.Factory
     interface Factory {
-        fun create(movieId: Int): MovieDetailViewModel
+        fun create(movieId: Long): MovieDetailViewModel
     }
 
     val loading = MutableLiveData<Boolean>()
